requirements.txt
Django==5.2
gunicorn==22.0.0
djangorestframework==3.15.1
pandas==2.2.0